move UFCParser logic to Flag
